+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/Dtos.java
@@ -2082,6 +2082,7 @@ public class Dtos {
     if (lastRun != null) {
       dto.setLastRunAlias(lastRun.getAlias());
       dto.setLastRunId(lastRun.getId());
+      dto.setLastRunInstrumentModelId(lastRun.getSequencer().getInstrumentModel().getId());
       dto.setLastSequencerId(lastRun.getSequencer().getId());
       dto.setLastSequencerName(lastRun.getSequencer().getName());
     }